requirements.txt
pytest
streamlit
pyyaml
